backend/lambda/codex/requirements.txt
openai>=1.0.0
boto3>=1.28.0
psycopg2-binary>=2.9.0

